Updating the functions to enable library logs.
The LibLogLevel attribute from OFSwitc13Device was removed, as it doesn’t make sense this as an attribute since the logging mechanism can’t be independeted by device (the library code is not objected-oriented). Now, a interface method is available to customize the logging system, and a simplified helper function (with the same name) can also be used to dump log messages to file.